requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distlib==0.3.9
Faker==33.0.0
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pillow==11.0.0
platformdirs==4.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
urwid==2.6.16
virtualenv==20.27.1
wcwidth==0.2.13
